+++ provision/core/phpcs/provision.sh
@@ -10,9 +10,14 @@ function php_codesniff_setup() {
   # Sniffs WordPress Coding Standards
   vvv_info " * Install/Update PHP_CodeSniffer (phpcs), see https://github.com/squizlabs/PHP_CodeSniffer"
   vvv_info " * Install/Update WordPress-Coding-Standards, sniffs for PHP_CodeSniffer, see https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards"
-  cd /srv/provision/phpcs
-  chown -R vagrant:www-data /srv/provision/phpcs
-  noroot composer update --no-ansi --no-autoloader --no-progress
+
+  if [[ -f "/srv/www/phpcs/CodeSniffer.conf" ]]; then
+    vvv_info " * Upgrading from old PHPCS setup"
+    rm -rf /srv/www/phpcs
+  fi
+  noroot mkdir -p /srv/www/phpcs
+  cd /srv/www/phpcs
+  COMPOSER_BIN_DIR="bin" noroot composer require --update-with-all-dependencies "dealerdirect/phpcodesniffer-composer-installer" "wp-coding-standards/wpcs" "automattic/vipwpcs" "phpcompatibility/php-compatibility" "phpcompatibility/phpcompatibility-paragonie" "phpcompatibility/phpcompatibility-wp" --no-ansi --no-progress
 
   # Link `phpcbf` and `phpcs` to the `/usr/local/bin` directory so
   # that it can be used on the host in an editor with matching rules
@@ -20,7 +25,6 @@ function php_codesniff_setup() {
   ln -sf "/srv/www/phpcs/bin/phpcs" "/usr/local/bin/phpcs"
 
   # Install the standards in PHPCS
-  noroot phpcs --config-set installed_paths ./CodeSniffer/Standards/WordPress/,./CodeSniffer/Standards/VIP-Coding-Standards/,./CodeSniffer/Standards/PHPCompatibility/,./CodeSniffer/Standards/PHPCompatibilityParagonie/,./CodeSniffer/Standards/PHPCompatibilityWP/
   noroot phpcs --config-set default_standard WordPress-Core
   noroot phpcs -i
 }